# rest-demo/modules/extensions/rest/gr-open-cosmote-rest-app-impl/rest-openapi.yaml
components:
    schemas:
        ProductPriceType:
            description: the product prive item type
            properties:
                preText:
                    type: string
                afterText:
                    type: string
                amount:
                    type: string
                currency:
                    type: string
        ProductPrice:
            description: the product price
            properties:
                monthlyPrice:
                    $ref: "#/components/schemas/ProductPriceType"
                onceOffPrice:
                    $ref: "#/components/schemas/ProductPriceType"
        ProductType:
            description: the product type
            properties:
                idRef:
                    type: string
                title:
                    type: string
        Product:
            description: the product
            properties:
                idRef:
                    type: string
                title:
                    type: string
                description:
                    type: string
                businessDescription:
                    type: string
                productType:
                    $ref: "#/components/schemas/ProductType"
                category:
                    type: string
                summary:
                    type: string
                businessSummary:
                    type: string
                imagePath:
                    type: string
                price:
                    $ref: "#/components/schemas/ProductPrice"
        Campaign:
            description: the campaign
            properties:
                address:
                    type: string
                comparisonFlag:
                    type: boolean
                eshopIdExisting:
                    type: string
                eshopIdTarget:
                    type: string
                specificPortalDescription:
                    type: string
                genevaBillingAccount:
                    type: string
                siebelProductIdTarget:
                    type: string
                siebelProductNameExisting:
                    type: string
                siebelProductNameTarget:
                    type: string
                campaignContactId:
                    type: string
                contactId:
                    type: string
                onceOffExisting:
                    type: string
                onceOfTarget:
                    type: string
                recurringExisting:
                    type: string
                recurringTarget:
                    type: string
                serialNum:
                    type: string
info:
    description: ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Campaigns"
    version: v1.0
openapi: 3.0.1
paths:
    "/campaigns/product/{idRef}":
        get:
            operationId: getCampaignProduct
            parameters:
                - in: path
                  name: idRef
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Product"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Product"
                    description: "Get Product"
            tags: ["Product"]
    "/campaigns/smspromo/{campaignContactId}/{contactId}":
        get:
            operationId: getCampaignSmspromoCampaignContactContactPage
            parameters:
                - in: path
                  name: campaignContactId
                  required: true
                  schema:
                      type: string
                - in: path
                  name: contactId
                  required: true
                  schema:
                      type: string
                - in: query
                  name: filter
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: search
                  schema:
                      type: string
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Campaign"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Campaign"
                                type: array
                    description: "Get Campaigns"
            tags: ["Campaign"]